exercises done

+++ part4/controllers/blogs.js
@@ -3,14 +3,15 @@ require('express-async-errors')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const userExtractor = require('../utils/middleware').userExtractor
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   }
+//   return null
+// }
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { name: 1, username: 1, _id: 1 })
@@ -20,7 +21,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
@@ -45,11 +46,26 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
   }
 
-
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog) {
+    return response.status(400).json({ error: 'invalid id' })
+  }
+
+  if(!(decodedToken && (decodedToken.id === blog.user.toString()))) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
+
+  const user = request.user
+
+  user.blogs = user.blogs.filter(blog => blog !== request.params.id)
+  await user.save()
   response.status(204).end()
 })
